+++ src/components/ModalBuild.tsx
@@ -14,6 +14,9 @@ import FormSelectBox from "./@shared/FormSelectBox";
 import FormTextBox from "./@shared/FormTextBox";
 import TextFieldAdd from "./TextFieldAdd";
 import TextFieldSaved from "./TextFieldSaved";
+import loginState from "../lib/recoil/auth";
+import { cloneUrlState } from "../lib/recoil/git";
+import buildOptionsState from "../lib/recoil/userBuildOptions";
 import useModal from "../lib/hooks/useModal";
 import useDeployEventHandler from "src/lib/hooks/useDeployEventHandler";
 import loginState from "../lib/recoil/auth";
