+++ src/components/ModalDeploy.tsx
@@ -15,7 +15,7 @@ import { ExpandMore as ExpandMoreIcon } from "@mui/icons-material";
 import Button from "./@shared/Button";
 import Config from "../lib/config";
 import useModal from "../lib/hooks/useModal";
-import { cloneRepoName } from "../lib/recoil/git/clone";
+import { cloneRepoName } from "../lib/recoil/git";
 
 function ModalDeploy() {
   const [socket, setSocket] = useState<Socket | null>(null);
