+++ src/components/ModalGlobal.tsx
@@ -8,6 +8,7 @@ import ModalDeleteConfirm from "./ModalDeleteConfirm";
 import { ModalCreateAlert, ModalDeleteAlert } from "./ModalAlert";
 import ModalPreview from "./ModalPreview";
 import ModalRepoDetails from "./ModalRepoDetails";
+import buildOptionsState from "../lib/recoil/userBuildOptions";
 import useModal from "../lib/hooks/useModal";
 
 import { modalState } from "../types/modal";
