+++ src/lib/recoil/git/atom.ts
@@ -1,0 +1,20 @@
+import { atom } from "recoil";
+
+import { GitNamespace, Repo } from "../../../types/projectOption";
+
+const cloneUrlState = atom({
+  key: "cloneUrlState",
+  default: "",
+});
+
+const gitNamespaceState = atom<GitNamespace[]>({
+  key: "gitNamespaceState",
+  default: [],
+});
+
+const gitRepoState = atom<Repo[]>({
+  key: "gitRepoState",
+  default: [],
+});
+
+export { cloneUrlState, gitNamespaceState, gitRepoState };
