--- src/lib/recoil/git/clone/atom.ts
@@ -1,8 +1,0 @@
-import { atom } from "recoil";
-
-const cloneUrlState = atom({
-  key: "cloneUrlState",
-  default: "",
-});
-
-export default cloneUrlState;
