--- src/lib/recoil/git/clone/index.ts
@@ -1,6 +1,0 @@
-import cloneUrlState from "./atom";
-import cloneRepoName from "./repoNameSelector";
-
-export { cloneRepoName };
-
-export default cloneUrlState;
