--- src/lib/recoil/git/namespace/atom.ts
@@ -1,10 +1,0 @@
-import { atom } from "recoil";
-
-import { GitNamespace } from "../../../../types";
-
-const gitNamespaceState = atom<GitNamespace[]>({
-  key: "gitNamespaceState",
-  default: [],
-});
-
-export default gitNamespaceState;
