--- src/lib/recoil/git/namespace/index.ts
@@ -1,6 +1,0 @@
-import gitNamespaceState from "./atom";
-import gitNamespaceList from "./namespaceSelector";
-
-export { gitNamespaceList };
-
-export default gitNamespaceState;
