--- src/lib/recoil/git/repos/atom.ts
@@ -1,10 +1,0 @@
-import { atom } from "recoil";
-
-import { Repo } from "../../../../types";
-
-const gitRepoState = atom<Repo[]>({
-  key: "gitRepoState",
-  default: [],
-});
-
-export default gitRepoState;
