--- src/lib/recoil/git/repos/index.ts
@@ -1,3 +1,0 @@
-import gitRepoState from "./atom";
-
-export default gitRepoState;
